pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-browser-agent"
version = "0.1.0"
description = "Visible browser AI agent with ref-based Playwright control, safety gates, compact context and transparent tool logs."
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "VLR test task implementation" }]
dependencies = [
  "anthropic>=0.49.0",
  "openai>=1.76.0",
  "playwright>=1.51.0",
  "pydantic>=2.10.0",
  "python-dotenv>=1.0.1",
  "rich>=13.9.0",
  "tenacity>=9.0.0",
  "tiktoken>=0.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.25.0",
]

[project.scripts]
ai-browser-agent = "ai_browser_agent.cli:main"

[tool.setuptools.packages.find]
include = ["ai_browser_agent*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

